Adds SPA support for GitHub Pages
Implements a workaround to support Single Page Applications (SPAs) when deployed to GitHub Pages.

This change introduces a 404.html page that redirects all requests to the index.html with the original URL encoded as a query parameter.
A script in index.html then parses this parameter and updates the browser's address bar using the History API, effectively simulating proper routing for the SPA.